review tweaks
