+++ actix-http/src/response.rs
@@ -198,7 +198,6 @@ impl<B> Response<B> {
 impl<B> fmt::Debug for Response<B>
 where
     B: MessageBody,
-    B::Error: Into<Error>,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let res = writeln!(
