+++ src/error/response_error.rs
@@ -12,90 +12,6 @@ use bytes::BytesMut;
 use crate::{__downcast_dyn, __downcast_get_type_id};
 use crate::{helpers, HttpResponse};
 
-/// General purpose actix web error.
-///
-/// An actix web error is used to carry errors from `std::error`
-/// through actix in a convenient way.  It can be created through
-/// converting errors with `into()`.
-///
-/// Whenever it is created from an external object a response error is created
-/// for it that can be used to create an HTTP response from it this means that
-/// if you have access to an actix `Error` you can always get a
-/// `ResponseError` reference from it.
-pub struct Error {
-    cause: Box<dyn ResponseError>,
-}
-
-impl Error {
-    /// Returns the reference to the underlying `ResponseError`.
-    pub fn as_response_error(&self) -> &dyn ResponseError {
-        self.cause.as_ref()
-    }
-
-    /// Similar to `as_response_error` but downcasts.
-    pub fn as_error<T: ResponseError + 'static>(&self) -> Option<&T> {
-        <dyn ResponseError>::downcast_ref(self.cause.as_ref())
-    }
-
-    /// TODO
-    pub fn error_response(&self) -> HttpResponse {
-        self.cause.error_response()
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Display::fmt(&self.cause, f)
-    }
-}
-
-impl fmt::Debug for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{:?}", &self.cause)
-    }
-}
-
-impl StdError for Error {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        None
-    }
-}
-
-impl From<std::convert::Infallible> for Error {
-    fn from(val: std::convert::Infallible) -> Self {
-        match val {}
-    }
-}
-
-/// `Error` for any error that implements `ResponseError`
-impl<T: ResponseError + 'static> From<T> for Error {
-    fn from(err: T) -> Error {
-        Error {
-            cause: Box::new(err),
-        }
-    }
-}
-
-impl From<Error> for Response<AnyBody> {
-    fn from(err: Error) -> Response<AnyBody> {
-        err.error_response().into()
-    }
-}
-
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-/////////////////////
-
 /// Errors that can generate responses.
 // TODO: add std::error::Error bound when replacement for Box<dyn Error> is found
 pub trait ResponseError: fmt::Debug + fmt::Display {
